+++ CHANGELOG.md
@@ -536,3 +536,11 @@
 * Sync the pin validation steps with other methods in import wallet
 * Updated tests wrt changes in vault generation and parameter validations
 * Implemented vault recovery using logs and updated tests
+
+### 1.30.05 (2023-09-08) Nightly version : only for experimental use
+
+* Implemented account recovery in case of deleted accounts, both generated and imported
+* Added test cases for account recovery
+* Added logs for export mnemonic
+* Added logs for export private key(for generated and imported accounts)
+* Re-encrypted the private key of imported wallets and updated vault state after change pin